Add quality check in command line
