+++ clinicadl/scripts/preprocessing/orig/run_qc_dl_OASIS.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-python /teams/ARAMIS/PROJECTS/junhao.wen/PhD/ADNI_classification/gitlabs/AD-DL/Code/image_preprocessing/quality_check_dl.py --caps_dir=/teams/ARAMIS/PROJECTS/CLINICA/CLINICA_datasets/temp/CAPS_OASIS_DL --tsv=/teams/ARAMIS/PROJECTS/junhao.wen/PhD/ADNI_classification/gitlabs/AD-DL/tsv_files/qc_tsvs/OASIS_all_subjects.tsv
+python /teams/ARAMIS/PROJECTS/junhao.wen/PhD/ADNI_classification/gitlabs/AD-DL/Code/image_preprocessing/quality_check.py --caps_dir=/teams/ARAMIS/PROJECTS/CLINICA/CLINICA_datasets/temp/CAPS_OASIS_DL --tsv=/teams/ARAMIS/PROJECTS/junhao.wen/PhD/ADNI_classification/gitlabs/AD-DL/tsv_files/qc_tsvs/OASIS_all_subjects.tsv
